Differences between Git and Github
Git is mainly concerned with Version Control. Git is a software that keeps track of all the history of all your code and changes made over time while working on, on your PC. It also has these features that enables collaboration and give freedom for two or more people to work on the same project together.
Git is actual version control software. While Github is a web service where we can have an account an apply or run Git in it. Github basically runs the Git software. So the projects that are done through Git can be run on Github in a collaborative way.
We can create repositories or projects or files. This can be done by uploading our files on to the repository. This displays all our code that we have been working on and others can see it or work on it. This really a great tool for teams in the company to effectively work on their projects. 
Above is the screen shot of the repositories I have created which displays all the uploaded files when clicked on them.